# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/beds": {
      "get": {
        "operationId": "BedsController_findAll",
        "summary": "Get all beds",
        "parameters": [
          {
            "name": "ward",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "available",
                "occupied",
                "critical",
                "maintenance"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of beds"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "BedsController_create",
        "summary": "Create a new bed record",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bed creation result (check success field)"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/beds/stats/overview": {
      "get": {
        "operationId": "BedsController_getStats",
        "summary": "Get bed statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Bed statistics retrieved"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/beds/ward/{ward}": {
      "get": {
        "operationId": "BedsController_findByWard",
        "summary": "Get beds by ward",
        "parameters": [
          {
            "name": "ward",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of ward beds retrieved"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/beds/available": {
      "get": {
        "operationId": "BedsController_getAvailableBeds",
        "summary": "Get all available beds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of available beds"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/beds/patient/{patient}": {
      "get": {
        "operationId": "BedsController_findByPatient",
        "summary": "Get bed by patient name",
        "parameters": [
          {
            "name": "patient",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bed retrieved"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/beds/occupancy": {
      "get": {
        "operationId": "BedsController_getOccupancyByWard",
        "summary": "Get bed occupancy statistics by ward",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Occupancy statistics retrieved"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/beds/{id}": {
      "get": {
        "operationId": "BedsController_findById",
        "summary": "Get bed by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bed details retrieved"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "put": {
        "operationId": "BedsController_update",
        "summary": "Update bed details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bed updated successfully"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "BedsController_patchUpdate",
        "summary": "Partially update bed details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bed updated successfully"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "BedsController_delete",
        "summary": "Delete a bed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bed deleted successfully"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/beds/{id}/allocate": {
      "patch": {
        "operationId": "BedsController_allocate",
        "summary": "Allocate a bed to a patient",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocateBedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bed allocated successfully"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/beds/{id}/release": {
      "patch": {
        "operationId": "BedsController_release",
        "summary": "Release a bed from a patient",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bed released successfully"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/beds/{id}/status": {
      "patch": {
        "operationId": "BedsController_updateStatus",
        "summary": "Update bed status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bed status updated successfully"
          }
        },
        "tags": [
          "Beds"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "NexCare Hospital Management API",
    "description": "Complete REST API for the NexCare Hospital Administrative Operations Platform. Features role-based access control (RBAC), JWT authentication, and comprehensive modules for patients, appointments, billing, beds, ambulance, inventory, feedback, and system management.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      },
      "x-user-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "User role for RBAC (superuser, administrative_staff, patient, ambulance, doctor, nurse)"
      }
    },
    "schemas": {
...
      "CreateBedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ICU-1",
            "description": "Unique identifier for the bed"
          },
          "ward": {
            "type": "string",
            "example": "ICU",
            "description": "Ward the bed belongs to"
          }
        },
        "required": [
          "id",
          "ward"
        ]
      },
      "UpdateBedDto": {
        "type": "object",
        "properties": {
          "ward": {
            "type": "string",
            "example": "ICU",
            "description": "Ward the bed belongs to"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "occupied",
              "critical",
              "maintenance"
            ],
            "example": "occupied",
            "description": "Status of the bed"
          },
          "patient": {
            "type": "string",
            "example": "John Doe",
            "description": "Patient name"
          }
        }
      },
      "AllocateBedDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "example": "P001",
            "description": "ID of the patient"
          },
          "admissionType": {
            "type": "string",
            "example": "Emergency",
            "description": "Type of admission"
          },
          "allocatedBy": {
            "type": "string",
            "example": "U002",
            "description": "Staff who allocated the bed"
          },
          "notes": {
            "type": "string",
            "example": "Patient needs monitoring",
            "description": "Additional notes"
          }
        },
        "required": [
          "patientId"
        ]
      },
...
    }
  }
}